% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params-docs.R
\name{date_param}
\alias{date_param}
\title{Date Parameter}
\arguments{
\item{date}{The date for which data is to be retrieved, as a character string
in "YYYY-MM-DD" format. The date should be based on Beijing Time (UTC+8).}
}
\description{
Date Parameter
}
