% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-company.R
\name{date_range_params}
\alias{date_range_params}
\title{Date Range Parameters}
\arguments{
\item{start_date}{The start date in "YYYY-MM-DD" format.}

\item{end_date}{The end date in "YYYY-MM-DD" format. If not provided, the
default value is last Monday. The interval between the start and end dates
must not exceed 10 years.}
}
\description{
Date Range Parameters
}
