% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params-docs.R
\name{limit_param}
\alias{limit_param}
\title{Limit Parameter}
\arguments{
\item{limit}{The number of latest records to return. This parameter is only
effective when requesting data within a date range.}
}
\description{
Limit Parameter
}
