% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lixingr2-package.R
\docType{package}
\name{lixingr2-package}
\alias{lixingr2}
\alias{lixingr2-package}
\title{lixingr2: R wrapper for Lixinger's Web API}
\description{
In the vision of its creator, Lixinger is both excellent code and product, designed to serve investors in Mainland China and Hong Kong by providing data services. It stands out for its competitive pricing compared to peers. This package encapsulates the Lixinger Open Platform API, facilitating the efficient writing of repeatable and thorough analytical reports.
}
\author{
\strong{Maintainer}: Changde Tan \email{tanchangde@gmail.com}

}
\keyword{internal}
