% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lxr_clear_token}
\alias{lxr_clear_token}
\title{Clear Lixinger API Token}
\usage{
lxr_clear_token()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Remove the stored API token from the environment variable.
}
\examples{
\dontrun{
lxr_clear_token()
}

}
