% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-company.R
\name{lxr_cn_com_fundmtl_security}
\alias{lxr_cn_com_fundmtl_security}
\title{Retrieve Fundamental Data of Listed Securities}
\usage{
lxr_cn_com_fundmtl_security(
  token,
  stock_codes,
  metrics_list,
  date = NULL,
  start_date = NULL,
  end_date = NULL,
  limit = NULL,
  .hdrs = list(),
  .config = list()
)
}
\arguments{
\item{token}{Authentication token, must be a character string.}

\item{stock_codes}{A vector of stock codes as character strings. Example format:
\code{c("300750", "600519", "600157")}.}

\item{metrics_list}{A vector of string metrics, e.g., \code{c("mc", "pe_ttm", "pb", "dyr")}.

Currently supported metrics:

\strong{Valuation Metrics}
\itemize{
\item PE-TTM : \code{pe_ttm}
\item PE-TTM (ex-NRI) : \code{d_pe_ttm}
\item PB : \code{pb}
\item PB (ex-Goodwill) : \code{pb_wo_gw}
\item PS-TTM : \code{ps_ttm}
\item Dividend Yield : \code{dyr}
\item PCF-TTM : \code{pcf_ttm}
\item Stock Price : \code{sp}
\item Price Change (\%) : \code{spc}
\item Intraday Range (\%) : \code{spa}
\item Trading Volume : \code{tv}
\item Trading Value : \code{ta}
\item Turnover Rate (\%) : \code{to_r}
\item Number of Shareholders : \code{shn}
\item Market Capitalization : \code{mc}
\item A-share Market Cap : \code{mc_om}
\item Tradable Market Cap : \code{cmc}
\item Free-float Market Cap : \code{ecmc}
\item Free-float Market Cap / Holder : \code{ecmc_psh}
\item Margin Purchase Value : \code{fpa}
\item Margin Repayment Value : \code{fra}
\item Margin Financing Balance : \code{fb}
\item Short-sale Value : \code{ssa}
\item Short-sale Repayment Value : \code{sra}
\item Short Interest Balance : \code{sb}
\item Northbound Connect Holdings (Shares) : \code{ha_sh}
\item Northbound Connect Holdings (Value)  : \code{ha_shm}
\item Northbound Connect Net Buy (Value)   : \code{mm_nba}
}

\strong{Valuation Statistics Metrics}

Format: \code{metricsName.granularity.statisticsDataType}

\emph{metricsName}
\itemize{
\item PE-TTM : \code{pe_ttm}
\item PE-TTM (ex-NRI) : \code{d_pe_ttm}
\item PB : \code{pb}
\item PB (ex-Goodwill) : \code{pb_wo_gw}
\item PS-TTM : \code{ps_ttm}
}

\emph{granularity}
\itemize{
\item Since Listing : \code{fs}
\item 20 Years : \code{y20}
\item 10 Years : \code{y10}
\item 5 Years : \code{y5}
\item 3 Years : \code{y3}
\item 1 Year : \code{y1}
}

\emph{statisticsDataType}
\itemize{
\item Percentile Position (\%) : \code{cvpos}
\item 20th-percentile Value : \code{q2v}
\item 50th-percentile Value : \code{q5v}
\item 80th-percentile Value : \code{q8v}
\item Minimum Value : \code{minv}
\item Maximum Value : \code{maxv}
\item Maximum Positive Value : \code{maxpv}
\item Average Value : \code{avgv}
}}

\item{date}{The date for which data is to be retrieved, as a character string
in "YYYY-MM-DD" format. The date should be based on Beijing Time (UTC+8).}

\item{start_date}{The start date as a character string in "YYYY-MM-DD" format.
The date should be based on Beijing Time (UTC+8).}

\item{end_date}{The end date as a character string in "YYYY-MM-DD" format.
If not provided, the default value is last Monday. The interval between
the start and end dates must not exceed 10 years. The date should be based
on Beijing Time (UTC+8).}

\item{limit}{The number of latest records to return.}

\item{.hdrs}{Additional headers to include in the request, passed as a named
list. Default is \code{list("Content-Type" = "application/json")}.}

\item{.config}{Additional configuration parameters, passed as a named list.
Can be partially or fully overridden.
Default is \code{list(max_tries = 4L, backoff_fun = NULL, retry_on = NULL,  return_format = "list", verbosity = NULL)}.

\code{max_tries}: Maximum number of retry attempts.

\code{backoff_fun}: Function to calculate backoff delay in seconds between retries.
Default is an exponential backoff with jitter. The delay increases with
the number of retry attempts and includes randomness to reduce conflicts.
The maximum delay is capped at 10 seconds.

\code{retry_on}: Predicate function that takes a single argument (the response)
and returns TRUE or FALSE to specify whether the response represents a
transient error that should trigger a retry. Default retries on HTTP status
codes 429 (Too Many Requests) and 500+ (Server Errors).

\code{return_format}:
\itemize{
\item \code{"json"}: Returns raw JSON data.
\item \code{"list"}: Returns data as an R list.
\item \code{"tibble"}: Returns data as a tibble (if implemented).
\item \code{"resp"}: Returns the performed request object.
}

\code{verbosity}: Controls the verbosity of printed information. This is a
wrapper around \code{\link[httr2:req_verbose]{httr2::req_verbose()}} that uses an integer to specify verbosity levels:
\itemize{
\item \code{0}: No output.
\item \code{1}: Show headers.
\item \code{2}: Show headers and bodies.
\item \code{3}: Show headers, bodies, and curl status messages.
}}
}
\value{
A parsed API response.
}
\description{
Retrieve Fundamental Data of Listed Securities
}
\details{
The length of \code{stock_codes} must be between 1 and 100. When assigning
a value to \code{start_date}, the length of \code{stock_codes} must be 1. Either
\code{start_date} or \code{date} must be provided. \code{limit} is only effective when
requesting data within a date range. When the length of \code{stock_codes} is
greater than 1, a maximum of 48 metrics can be selected; when the length of
\code{stock_codes} equals 1, a maximum of 36 metrics can be retrieved.
}
\examples{
\dontrun{
lxr_cn_com_fundmtl_security(
  token = Sys.getenv("LIXINGR_TOKEN"),
  start_date = "2025-07-01",
  end_date = "2025-07-15",
  stock_codes = "600030",,
  metrics_list = c("pe_ttm", "mc", "pe_ttm.y5.cvpos")
)
}

}
\seealso{
\href{https://www.lixinger.com/open/api/doc?api-key=cn/company/fundamental/security}{API documentation}
}
