% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-company.R
\name{lxr_cn_com_profile}
\alias{lxr_cn_com_profile}
\title{Get Company Profile Data}
\usage{
lxr_cn_com_profile(token, stock_codes, .hdrs = list(), .config = list())
}
\arguments{
\item{token}{Authentication token, must be a character string.}

\item{stock_codes}{A vector of stock codes as character strings. Example format:
\code{c("300750", "600519", "600157")}.}

\item{.hdrs}{Additional headers to include in the request, passed as a named
list. Default is \code{list("Content-Type" = "application/json")}.}

\item{.config}{Additional configuration parameters, passed as a named list.
Can be partially or fully overridden.
Default is \code{list(max_tries = 4L, backoff_fun = NULL, retry_on = NULL,  return_format = "list", verbosity = NULL)}.

\code{max_tries}: Maximum number of retry attempts.

\code{backoff_fun}: Function to calculate backoff delay in seconds between retries.
Default is an exponential backoff with jitter. The delay increases with
the number of retry attempts and includes randomness to reduce conflicts.
The maximum delay is capped at 10 seconds.

\code{retry_on}: Predicate function that takes a single argument (the response)
and returns TRUE or FALSE to specify whether the response represents a
transient error that should trigger a retry. Default retries on HTTP status
codes 429 (Too Many Requests) and 500+ (Server Errors).

\code{return_format}:
\itemize{
\item \code{"json"}: Returns raw JSON data.
\item \code{"list"}: Returns data as an R list.
\item \code{"tibble"}: Returns data as a tibble (if implemented).
\item \code{"resp"}: Returns the performed request object.
}

\code{verbosity}: Controls the verbosity of printed information. This is a
wrapper around \code{\link[httr2:req_verbose]{httr2::req_verbose()}} that uses an integer to specify verbosity levels:
\itemize{
\item \code{0}: No output.
\item \code{1}: Show headers.
\item \code{2}: Show headers and bodies.
\item \code{3}: Show headers, bodies, and curl status messages.
}}
}
\value{
A parsed API response.
}
\description{
Get Company Profile Data
}
\details{
The length of \code{stock_codes} must be between 1 and 100.
}
\seealso{
\href{https://www.lixinger.com/open/api/doc?api-key=cn/company/profile}{API documentation}
}
