% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lxr_get_token}
\alias{lxr_get_token}
\title{Get Lixinger API Token}
\usage{
lxr_get_token()
}
\value{
Character string. The API token.
}
\description{
Retrieve the Lixinger API token from the environment variable.
}
\examples{
\dontrun{
token <- lxr_get_token()
}

}
\seealso{
\code{\link[=lxr_set_token]{lxr_set_token()}} for setting the token.
}
