% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lxr_has_token}
\alias{lxr_has_token}
\title{Check if Token is Set}
\usage{
lxr_has_token()
}
\value{
Logical. \code{TRUE} if token is set, \code{FALSE} otherwise.
}
\description{
Check whether a Lixinger API token is available.
}
\examples{
lxr_has_token()

}
