% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-company.R
\name{lxr_hk_com}
\alias{lxr_hk_com}
\title{Get Detailed Stock Information}
\usage{
lxr_hk_com(
  token,
  stock_codes = NULL,
  fs_table_type = NULL,
  mutual_markets = NULL,
  include_delisted = NULL,
  .hdrs = list(),
  .config = list()
)
}
\arguments{
\item{token}{Authentication token, must be a character string.}

\item{stock_codes}{A vector of stock codes as character strings.}

\item{fs_table_type}{Financial statement type as a character string. Currently
supported types include: \code{"non_financial"}, \code{"bank"}, \code{"insurance"},
\code{"security"}, \code{"reit"} and \code{"other_financial"}.}

\item{mutual_markets}{Stock Connect type as a character string. Currently
supports \code{"ah"} (Shanghai-Hong Kong Stock Connect).}

\item{include_delisted}{Whether to include delisted stocks. \code{FALSE} or
\code{NULL} means not included.}

\item{.hdrs}{Additional headers to include in the request, passed as a named
list. Default is \code{list("Content-Type" = "application/json")}.}

\item{.config}{Additional configuration parameters, passed as a named list.
Can be partially or fully overridden.
Default is \code{list(max_tries = 4L, backoff_fun = NULL, retry_on = NULL,  return_format = "list", verbosity = NULL)}.

\code{max_tries}: Maximum number of retry attempts.

\code{backoff_fun}: Function to calculate backoff delay in seconds between retries.
Default is an exponential backoff with jitter. The delay increases with
the number of retry attempts and includes randomness to reduce conflicts.
The maximum delay is capped at 10 seconds.

\code{retry_on}: Predicate function that takes a single argument (the response)
and returns TRUE or FALSE to specify whether the response represents a
transient error that should trigger a retry. Default retries on HTTP status
codes 429 (Too Many Requests) and 500+ (Server Errors).

\code{return_format}:
\itemize{
\item \code{"json"}: Returns raw JSON data.
\item \code{"list"}: Returns data as an R list.
\item \code{"tibble"}: Returns data as a tibble (experimental).
\item \code{"resp"}: Returns the performed request object.
}

\code{verbosity}: Controls the verbosity of printed information. This is a
wrapper around \code{\link[httr2:req_verbose]{httr2::req_verbose()}} that uses an integer to specify verbosity levels:
\itemize{
\item \code{0}: No output.
\item \code{1}: Show headers.
\item \code{2}: Show headers and bodies.
\item \code{3}: Show headers, bodies, and curl status messages.
}}
}
\value{
A parsed API response.
}
\description{
Get Detailed Stock Information
}
\details{
If \code{stock_codes} is not provided, all stock codes will be queried.
}
\seealso{
\href{https://www.lixinger.com/open/api/doc?api-key=hk/company}{API documentation}
}
