% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-company.R
\name{lxr_hk_com_fs_insurance}
\alias{lxr_hk_com_fs_insurance}
\title{Retrieve Financial Statement Data for Listed Insurance Companies}
\usage{
lxr_hk_com_fs_insurance(
  token,
  stock_codes,
  metrics_list,
  date = NULL,
  start_date = NULL,
  end_date = NULL,
  limit = NULL,
  options = list()
)
}
\arguments{
\item{token}{Authentication token, must be a character string.}

\item{stock_codes}{A vector of stock codes as character strings.}

\item{metrics_list}{A vector of metric names to retrieve. Refer to the API
documentation (see \strong{See Also} section) for the full list of supported
metrics for this endpoint.}

\item{date}{The date for which data is to be retrieved, as a character string
in "YYYY-MM-DD" format. The date should be based on Beijing Time (UTC+8).}

\item{start_date}{The start date as a character string in "YYYY-MM-DD" format.
The date should be based on Beijing Time (UTC+8).}

\item{end_date}{The end date as a character string in "YYYY-MM-DD" format.
If not provided, the default value is last Monday. The interval between
the start and end dates must not exceed 10 years. The date should be based
on Beijing Time (UTC+8).}

\item{limit}{The number of latest records to return.}

\item{options}{A list for advanced control over the request.
\itemize{
\item \code{extra_headers}: Additional headers to include in the request, passed as a
named list. Default is \code{list("Content-Type" = "application/json")}.
\item \code{call_config}: Additional configuration parameters, passed as a named list.
Can be partially or fully overridden.
Default is \code{list(max_tries = 4L, backoff_fun = NULL, retry_on = NULL, return_format = "list", verbosity = NULL)}.
\itemize{
\item \code{max_tries}: Maximum number of retry attempts.
\item \code{backoff_fun}: Function to calculate backoff delay in seconds between retries.
Default is an exponential backoff with jitter. The delay increases with
the number of retry attempts and includes randomness to reduce conflicts.
The maximum delay is capped at 10 seconds.
\item \code{retry_on}: Predicate function that takes a single argument (the response)
and returns TRUE or FALSE to specify whether the response represents a
transient error that should trigger a retry. Default retries on HTTP status
codes 429 (Too Many Requests) and 500+ (Server Errors).
\item \code{return_format}: Output format options:
\itemize{
\item \code{"json"}: Returns raw JSON data.
\item \code{"list"}: Returns data as an R list.
\item \code{"tibble"}: Returns data as a tibble (experimental).
\item \code{"resp"}: Returns the performed request object.
}
\item \code{verbosity}: Controls the verbosity of printed information. This is a
wrapper around \code{\link[httr2:req_verbose]{httr2::req_verbose()}} that uses an integer to specify verbosity levels:
\itemize{
\item \code{0}: No output.
\item \code{1}: Show headers.
\item \code{2}: Show headers and bodies.
\item \code{3}: Show headers, bodies, and curl status messages.
}
}
}}
}
\value{
A parsed API response.
}
\description{
Retrieve Financial Statement Data for Listed Insurance Companies
}
\details{
The length of \code{stock_codes} must be between 1 and 100. When assigning
a value to \code{start_date}, the length of \code{stock_codes} must be 1. Either
\code{start_date} or \code{date} must be provided. Passing \code{latest} as the value for
\code{date} will retrieve the most recent financial statement data within the
last 1.1 years.\code{limit} is only effective when requesting data within a date
range. When the length of \code{stock_codes} is greater than 1, a maximum of 48
metrics can be selected; when the length of \code{stock_codes} equals 1,
a maximum of 128 metrics can be retrieved.
}
\examples{
\dontrun{
lxr_hk_com_fs_insurance(
  start_date = "2024-07-01",
  end_date = "2025-07-01",
  stock_codes = "01336",
  metrics_list = "q.ps.oi.t"
)
}

}
\seealso{
\href{https://www.lixinger.com/open/api/doc?api-key=hk/company/fs/insurance}{API documentation}
}
