% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-factory.R
\name{make_endpoint}
\alias{make_endpoint}
\title{Endpoint factory}
\usage{
make_endpoint(client = new_client(), endpoint, required, optional = NULL)
}
\arguments{
\item{client}{An API client.}

\item{endpoint}{Character. API path relative to \code{base_url}.}

\item{required}{Character vector. Names of parameters \strong{required} by this endpoint.}

\item{optional}{Character vector. Names of optional parameters.}
}
\value{
A user-callable function.
}
\description{
Endpoint factory
}
