% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params-docs.R
\name{metrics_list_param}
\alias{metrics_list_param}
\title{Metrics List Parameter}
\arguments{
\item{metrics_list}{A vector of metric names to retrieve. Refer to the API
documentation (see \strong{See Also} section) for the full list of supported
metrics for this endpoint.}
}
\description{
Metrics List Parameter
}
\keyword{internal}
