% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params-docs.R
\name{stock_code_param}
\alias{stock_code_param}
\title{Stock Code Parameter}
\arguments{
\item{stock_code}{A single stock code as a character string.}
}
\description{
Stock Code Parameter
}
\keyword{internal}
