% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params-docs.R
\name{stock_codes_param}
\alias{stock_codes_param}
\title{Stock Codes Parameter}
\arguments{
\item{stock_codes}{A vector of stock codes as character strings.}
}
\description{
Stock Codes Parameter
}
\keyword{internal}
