% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params-docs.R
\name{common_return}
\alias{common_return}
\title{Common Return Descriptions}
\value{
A parsed API response. The format of the returned data depends on the \code{.return_format} parameter:
\itemize{
\item \code{"json"}: Returns raw JSON data.
\item \code{"list"}: Returns data as an R list.
\item \code{"tibble"}: Returns data as a tibble (if implemented).
}
}
\description{
Common Return Descriptions
}
