% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-company.R
\name{date_range_params}
\alias{date_range_params}
\title{Date Range Parameters}
\arguments{
\item{start_date}{The start date as a character string in "YYYY-MM-DD" format.
The date should be based on Beijing Time (UTC+8).}

\item{end_date}{The end date as a character string in "YYYY-MM-DD" format.
If not provided, the default value is last Monday. The interval between
the start and end dates must not exceed 10 years. The date should be based
on Beijing Time (UTC+8).}
}
\description{
Date Range Parameters
}
