% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-factory.R
\name{make_endpoint}
\alias{make_endpoint}
\title{Generate API endpoint function}
\usage{
make_endpoint(endpoint, required, optional = NULL)
}
\arguments{
\item{endpoint}{API path, the personalized part of the API path}

\item{required}{Character vector of required parameter names}

\item{optional}{Character vector of optional parameter names}
}
\value{
Returns an API endpoint function
}
\description{
Generate API endpoint function
}
